Add exception for when the FONA is in data mode
If the FONA is stuck in data mode, the chatscript will repeatedly fail with 12 second waits for the FONA to reply to 'AT'. By sending +++ on failure, it at least gives the FONA a chance to switch between command and data mode a few times before it's back to a consistent state.